+++ apps/newsletters-ui/src/app/components/SchemaForm/SchemaField.tsx
@@ -1,10 +1,8 @@
 import type { z } from 'zod';
-import {
-	BooleanInput,
-	NumberInput,
-	OptionalNumberInput,
-	SelectInput,
-} from './formControls';
+import { BooleanInput } from "./BooleanInput";
+import { NumberInput } from "./NumberInput";
+import { OptionalNumberInput } from "./OptionalNumberInput";
+import { SelectInput } from "./SelectInput";
 import { StringInput } from "./StringInput";
 import type { FieldDef, FieldValue, NumberInputSettings } from './util';
 import { fieldValueAsDisplayString } from './util';
